Fix namespace for types in Cake.GitLabCI.Module
When the GitLab CI module was introduced initially, the types were placed in the 'Cake.AzurePipelines.Module' namespace by accident.
To fix this and make the namespaces consistent with the package and assembly structure, move all types to the 'Cake.GitLabCI.Module' namespace.

While this is a breaking change, the impact is mitigated in two ways:

- Builds using the Cake scripting runner automatically discover modules by looking for the 'CakeModule' attribute. The change should thus be transparent to these projects
- For Cake.Frosting based builds, a second "GitLabCIModule" type is kept in the Cake.AzurePipelines.Module namespace so the module should be source-compatible with earlier versions

The GitLabCIModule in Cake.AzurePipelines.Module was marked as obsolete to discourage usage.