(#306) Add GitHub Report Printer
This will be used to improve the Cake output when running on GitHub
Actions.  This includes the task summary being written out to the Step
Summary.

With the latest releases of Cake, the task summary is written out even
when there is a failed build, so this means that we can capture that
output into the Step Summary, to get an overview of which task might
have failed the build.